App/requirements.txt

###### Requirements with Version Specifiers ######
numpy >= 1.17.3
matplotlib == 3.2.2
pillow == 9.0.1
tensorflow == 2.8.0
keras == 2.8.0
